+++ src/main/java/roomescape/domain/theme/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.domain.Theme;
+package roomescape.domain.theme;
 
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
