+++ src/main/java/roomescape/domain/theme/ThemeName.java
@@ -1,4 +1,4 @@
-package roomescape.domain.Theme;
+package roomescape.domain.theme;
 
 public class ThemeName {
