+++ src/main/java/roomescape/domain/theme/Thumbnail.java
@@ -1,4 +1,4 @@
-package roomescape.domain.Theme;
+package roomescape.domain.theme;
 
 public class Thumbnail {
