+++ src/main/java/roomescape/dto/reservation/ReservationRequest.java
@@ -3,10 +3,10 @@ package roomescape.dto.reservation;
 import java.time.LocalDate;
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.domain.Reservation.Reservation;
-import roomescape.domain.ReservationTime.ReservationTime;
-import roomescape.domain.Theme.Theme;
 import roomescape.domain.member.Member;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.time.ReservationTime;
 
 public record ReservationRequest(
         @JsonFormat(pattern = "yyyy-MM-dd") LocalDate date,
