+++ src/main/java/roomescape/dto/reservation/ReservationResponse.java
@@ -2,10 +2,10 @@ package roomescape.dto.reservation;
 
 import java.time.LocalDate;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.domain.Reservation.Reservation;
-import roomescape.domain.ReservationTime.ReservationTime;
-import roomescape.domain.Theme.Theme;
 import roomescape.domain.member.Member;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.time.ReservationTime;
 import roomescape.dto.member.MemberResponse;
 import roomescape.dto.reservationtime.ReservationTimeResponse;
 import roomescape.dto.theme.ThemeResponse;
