+++ src/main/java/roomescape/dto/theme/ThemeRequest.java
@@ -1,7 +1,7 @@
 package roomescape.dto.theme;
 
 import java.util.Objects;
-import roomescape.domain.Theme.Theme;
+import roomescape.domain.theme.Theme;
 
 public record ThemeRequest(
         String name,
