+++ src/main/java/roomescape/dto/theme/ThemeResponse.java
@@ -1,6 +1,6 @@
 package roomescape.dto.theme;
 
-import roomescape.domain.Theme.Theme;
+import roomescape.domain.theme.Theme;
 
 public record ThemeResponse(
         Long id,
