+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -3,7 +3,7 @@ package roomescape.repository;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.ReservationTime.ReservationTime;
+import roomescape.domain.time.ReservationTime;
 
 public interface ReservationTimeRepository {
