+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -2,7 +2,7 @@ package roomescape.repository;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.Theme.Theme;
+import roomescape.domain.theme.Theme;
 
 public interface ThemeRepository {
