+++ src/main/java/roomescape/repository/implementation/JdbcReservationRepository.java
@@ -13,11 +13,11 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation.Reservation;
-import roomescape.domain.ReservationTime.ReservationTime;
-import roomescape.domain.Theme.Theme;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.time.ReservationTime;
 import roomescape.repository.ReservationRepository;
 
 @Repository
