+++ src/main/java/roomescape/repository/implementation/JdbcThemeRepository.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Theme.Theme;
+import roomescape.domain.theme.Theme;
 import roomescape.repository.ThemeRepository;
 
 @Repository
